// src/app/services/models/product.ts
export class Product {
  _id?: string;
  product_name?: string;
  brand_name?: string;
  description?: string;
  allergens?: string;
  nutrition_facts?: any;
  price?: any;
  expiration_date?: string;
  country_of_origin?: string;
  stocks?: any;
  barcode?: string;
  product_category?: string;
  owner?: any;
  product_image?: any;
  buyers?: any[];
  createdAt?: any;
  updatedAt?: any;
}